// src/types/recentFileType.ts
import { ICreateByTypes } from "./filesType";

export interface IRecentTypes {
  actionDate: Date;
  actionStatus: string;
  aproveDownloadPublic: number;
  checkFile: string;
  createdAt: Date;
  createdBy: ICreateByTypes;
  detail: string;
  device: string;
  downloadBy: ICreateByTypes;
  downloadStatus: string;
  dropExpiredAt: string;
  dropLink: string;
  dropStatus: string;
  dropUrl: string;
  expired: string;
  favorite: 1;
  filePassword: string;
  fileType: string;
  filename: string;
  folder_id: { _id: string };
  getLinkBy: number;
  id: string;
  ip: string;
  isDeleted: number;
  isPublic: string;
  longUrl: string;
  newFilename: string;
  newPath: string;
  passwordUrlAll: string;
  path: string;
  permissionSharePublic: 0;
  restore: string;
  shortUrl: string;
  size: string;
  source: string;
  status: string;
  totalDownload: 0;
  totalDownloadFaild: 0;
  type_id: { _id: string };
  updatedAt: Date;
  updatedBy: ICreateByTypes;
  uploadBy: { _id: string };
  uploadStatus: string;
  url: string;
  urlAll: string;
  _id: string;
}